refactor: 컨벤션에 맞게 파일 합체
